refactor(codeguru-reviewer): convert to libnuke resource standard format
